Modularice el codigo del carrito, detalles de css
