+++ README.md
@@ -6,4 +6,12 @@ Franco Gonzalez Donato
 
 Suleivi Liendo
 
-Gabriela Yael Moreno Polacov
+Gabriela Yael Moreno Polacov:
+Github: gabipolacov
+- Arme el grid de las cards, el header y el footer
+- Traje los productos de la API (api.js)
+- Hice el modal de detalle de cada producto (detalle.js)
+- Agregué Filtros al buscador
+- Agregué la funcion que permite al usuario escribir mientras se filtran los elementos
+- Detalles de CSS para los anteriores
+- Fui modularizando el código en los distintos componentes
