+++ index.html
@@ -23,7 +23,6 @@
 
 <body>
     <header>
-
         <!--Menú de navegación-->
         <nav class="navbar bg-body-tertiary fixed-top" id="nav-bar">
             <div class="container-fluid" id="content-nav">
@@ -34,7 +33,8 @@
                 <div class="botones-nav">
                     <button id="openCartBtn" class="btn btn-primary" type="button" data-bs-toggle="offcanvas"
                         data-bs-target="#carritoDeCompras" aria-controls="carritoDeCompras">
-                        <i class="fa-solid fa-cart-shopping"></i> <span id="cartCountSpan" class="badge bg-secondary">0</span>
+                        <i class="fa-solid fa-cart-shopping"></i> <span id="cartCountSpan"
+                            class="badge bg-secondary">0</span>
                     </button>
 
                     <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas"
@@ -42,12 +42,11 @@
                         id="btnMenu">
                         <span class="navbar-toggler-icon"></span>
                     </button>
-
                 </div>
             </div>
         </nav>
-        <!--Sidebar de carrito de compras-->
 
+        <!--Sidebar de carrito de compras-->
         <div class="offcanvas offcanvas-end" tabindex="-1" id="carritoDeCompras" aria-labelledby="carritoDeCompras">
             <div class="offcanvas-header">
                 <h5 class="carritoDeCompras-title" id="offcanvasRightLabel">Carro de compras</h5>
@@ -98,9 +97,6 @@
                 </ul>
 
             </div>
-        </div>
-        </div>
-        </nav>
 
     </header>
 
@@ -147,22 +143,37 @@
         </div>
 
         <!-- Modal de detalle -->
-        <div class="modal fade" id="productDetailModal" tabindex="-1" aria-labelledby="productModalLabel" aria-hidden="true">
+        <div class="modal fade" id="productDetailModal" tabindex="-1" aria-labelledby="productModalLabel"
+            aria-hidden="true">
             <!-- Generado en JS -->
         </div>
         <div id="addToCartMessage" class="add-to-cart-message"></div>
 
         <!-- Toast de Agregado al Carrito -->
         <div class="toast-container position-fixed bottom-0 end-0 p-3">
-        <div id="liveToast" class="toast" role="alert" aria-live="assertive" aria-atomic="true">
-            <div class="toast-header">
-            <strong class="me-auto">Etna Moda</strong>
-            <button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
-            </div>
-            <div class="toast-body">
-            Item agregado al carrito!
+            <div id="liveToast" class="toast" role="alert" aria-live="assertive" aria-atomic="true">
+                <div class="toast-header">
+                    <strong class="me-auto">Etna Moda</strong>
+                    <button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
+                </div>
+                <div class="toast-body">
+                    Item agregado al carrito!
+                </div>
             </div>
         </div>
+        <!-- Toast de Compra Completada -->
+        <div id="purchaseCompleted" class="purchase-completed"></div>
+
+        <div class="toast-container position-fixed bottom-0 end-0 p-3">
+            <div id="liveToast1" class="toast" role="alert" aria-live="assertive" aria-atomic="true">
+                <div class="toast-header">
+                    <strong class="me-auto">Etna Moda</strong>
+                    <button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
+                </div>
+                <div class="toast-body">
+                    Compra Completada!
+                </div>
+            </div>
         </div>
 
     </main>
