+++ src/services/api.js
@@ -4,6 +4,6 @@ export function getProducts() {
   let resp = fetch('https://fakestoreapi.com/products')
     .then(response => response.json())
     .then((data) => data);
-  console.log(resp);
+
   return resp;
 }
