+++ style/style.css
@@ -78,17 +78,30 @@ main {
     border-color: rgb(255, 255, 255);
 }
 
-.item-cart-wrapper{
+.btn-primary {
+    --bs-btn-color: #fff;
+    --bs-btn-bg: none;
+    --bs-btn-border-color: #ffffff;
+    --bs-btn-hover-color: rgb(147, 47, 235);
+    --bs-btn-hover-bg: #ffffff;
+    --bs-btn-hover-border-color: rgb(147, 47, 235);
+}
+
+.item-cart-wrapper {
     display: grid;
     grid-template-columns: 1fr 7fr 1fr 1fr;
     align-items: center;
+    margin: 25px;
 }
 
-.remove-item-btn{
+.remove-item-btn {
     text-decoration: none;
     font-size: 10px;
 }
-.remove-item-btn:hover{
+
+
+
+.remove-item-btn:hover {
     opacity: 0.7;
 }
 
@@ -238,6 +251,15 @@ main {
     border-bottom: none;
 }
 
+.img-fluid {
+    justify-content: center;
+
+}
+
+.d-flex-1 {
+    display: flex;
+}
+
 
 .modal-footer {
     justify-content: center;
@@ -372,9 +394,8 @@ footer p {
     }
 
     footer .redes {
-        flex-direction: column;
-        column-gap: 0;
-        row-gap: 10px;
+        flex-direction: row;
+        column-gap: 70px;
         width: 100%;
     }
